requirements.txt
-i https://pypi.org/simple
blinker==1.4
certifi==2019.9.11
click==7.0
elastic-apm[flask]==5.3.2
flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
pika==1.1.0
python-decouple==3.3
python-logging-rabbitmq==2.0.0
urllib3==1.25.7
werkzeug==0.16.0
